allow image auto aspect ratio for FPGA GUI

+++ src/OneWare.UniversalFpgaProjectSystem/ViewModels/FpgaGuiElements/FpgaGuiElementCruviHsViewModel.cs
@@ -5,7 +5,7 @@ namespace OneWare.UniversalFpgaProjectSystem.ViewModels.FpgaGuiElements;
 
 public class FpgaGuiElementCruviHsViewModel : FpgaGuiElementInterfaceViewModelBase
 {
-    public FpgaGuiElementCruviHsViewModel(int x, int y) : base(x, y)
+    public FpgaGuiElementCruviHsViewModel(double x, double y) : base(x, y)
     {
     }
 }
