+++ src/OneWare.UniversalFpgaProjectSystem/ViewModels/FpgaGuiElements/FpgaGuiElementViewModelBase.cs
@@ -7,13 +7,13 @@ public abstract class FpgaGuiElementViewModelBase : ObservableObject
 {
     public IHardwareModel? Parent { get; init; }
     
-    public int X { get; }
+    public double X { get; }
     
-    public int Y { get; }
+    public double Y { get; }
 
     public double Rotation { get; init; }
 
-    public FpgaGuiElementViewModelBase(int x, int y)
+    public FpgaGuiElementViewModelBase(double x, double y)
     {
         X = x;
         Y = y;
